% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deltamethod_from_model.R
\name{deltamethod_from_model}
\alias{deltamethod_from_model}
\title{Predict the standard error from a fitted model using the delta method}
\usage{
deltamethod_from_model(model, tau, naive = FALSE)
}
\arguments{
\item{model}{A fitted model object. Supported types are
\code{\link[survival:survreg]{survival::survreg()}} with log-logistic distribution and
\code{\link[mets:binreg]{mets::logitIPCW()}}.}

\item{tau}{Numeric scalar. The time horizon at which the survival
probability is to be estimated.}

\item{naive}{Logical. If \code{TRUE}, use the naive variance
estimator for \code{binreg} models. If \code{FALSE} (default),
use the robust variance estimator.}
}
\value{
A \verb{function(df, z = 1.96)} that takes a data frame of covariates
\code{df} and returns a data frame with columns for the
prediction \code{prediction}, lower \code{lower} and upper \code{upper} Wald
confidence intervals, and standard error \code{se}. The function can also take
an optional argument \code{z} for the z-score used in confidence
interval calculation (default is 1.96 for 95\% confidence intervals).
}
\description{
\loadmathjax
Constructs a prediction function that estimates the survival probability
at a specified time horizon \mjeqn{\tau}{tau} from a fitted model,
and computes its standard error and Wald confidence interval
using the delta method.
}
\details{
The function supports models of from \code{\link[survival:survreg]{survival::survreg()}}
log-logistic distribution and models of class \code{binreg}
(such as those fitted by \code{\link[mets:binreg]{mets::logitIPCW()}})
\insertCite{Blanche2023,mets1,mets2}{IPCWJK}. For \code{binreg} models,
the function can use either the naive or the robust variance estimator,
depending on the value of the \code{naive} argument.
}
\examples{
library(survival)
tau <- 100
df <- veteran[, c("time", "status", "trt")]
newdata <- data.frame(trt = c(1, 2))

# Fit a log-logistic survival model
survreg_fit <- survreg(Surv(time, status) ~ trt,
  data = df,
  dist = "loglogistic"
)
pred_fun <- deltamethod_from_model(survreg_fit, tau = tau)
pred_fun(newdata)

# Fit a logitIPCW model
library(mets)
logipcw_fit <- logitIPCW(Event(time, status) ~ trt, time = tau, data = df)
predfun_logit <- deltamethod_from_model(logipcw_fit, tau = tau)
pred_fun(newdata)

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link[=deltamethod_pred_function]{deltamethod_pred_function()}} for the underlying implementation and
\link{IPCWJK} for confidence interval calculation
}
