% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipcw_logistic_regression.R
\name{ipcw_logistic_regression}
\alias{ipcw_logistic_regression}
\title{IPCW Logistic Regression}
\usage{
ipcw_logistic_regression(data, tau, time_var = "t", status_var = "delta")
}
\arguments{
\item{data}{A data frame containing the survival data. Must include columns
for the observed time and event indicator.}

\item{tau}{Numeric scalar. The time horizon at which the survival
probability is to be estimated.}

\item{time_var}{Character. The name of the variable in \code{data}
representing the observed time to event or censoring.
Default is \code{"t"}.}

\item{status_var}{Character. The name of the variable in \code{data}
representing the event indicator (1 if event occurred, 0 if censored).
Default is \code{"delta"}.}
}
\value{
An object of class \code{IPCWModel}.
}
\description{
\loadmathjax
Fits a logistic regression model with IPCW for
right-censored survival data. Jackknife model training is performed.
}
\details{
Training is performed using the \code{glm} function from the \code{stats} package,
using a quasibinomial family to account for the weights.
}
\examples{
#' # veteran data example
library(survival)
tau <- 80
df <- veteran[, c("time", "status", "trt")]
newdata <- data.frame(trt = c(1, 2))

fit <- ipcw_logistic_regression(df,
  tau = tau, time_var = "time",
  status_var = "status"
)
predict(fit, newdata)
}
\references{
\insertAllCited{}
}
\seealso{
Other ipcwmodels: 
\code{\link{ipcw_xgboost}()}
}
\concept{ipcwmodels}
