% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipcw_weights.R
\name{ipcw_weights}
\alias{ipcw_weights}
\title{Use a Kaplan-Meier estimator to estimate IPCW weights}
\usage{
ipcw_weights(data, tau, time_var = "t", status_var = "delta")
}
\arguments{
\item{data}{A data frame containing the survival data. Must include columns
for the observed time and event indicator.}

\item{tau}{Numeric scalar. The time horizon at which the survival
probability is to be estimated.}

\item{time_var}{Character. The name of the variable in \code{data}
representing the observed time to event or censoring.
Default is \code{"t"}.}

\item{status_var}{Character. The name of the variable in \code{data}
representing the event indicator (1 if event occurred, 0 if censored).
Default is \code{"delta"}.}
}
\value{
A numeric vector of IPCW weights, ordered as in the original
\code{data}. The weights are not normalized (do not sum to one).
}
\description{
\loadmathjax
Computes inverse probability of censoring weights (IPCW) for right-censored
survival data using the \code{\link[pec:ipcw]{pec::ipcw()}} function based on a Kaplan-Meier
estimate of the censoring distribution \mjeqn{C}{C}.
}
\details{
Weighting observations by their Inverse Probability of Censoring has been
proposed to prevent bias introduced by removing censored individuals
\insertCite{Vock2016,Ginestet2021,Blanche2023}{IPCWJK} and improve model
performance.

\mjdeqn{
\tilde{w}_i =
\left\lbrace
\begin{array}{ll}
 0, & c_i < \tau \wedge t^*_i \cr
 \frac{1}{P(C > \tau \mid X = x_i)}, & \tau < c_i \wedge t^*_i \cr
 \frac{1}{P(C > t_i \mid X = x_i)}, & t^*_i < c_i \wedge \tau
\end{array}
\right.
}{w_i_tilde = 0, if c_i < tau and t*_i; 1 / P(C > tau | X = x_i),
if tau < c_i and t*_i; 1 / P(C > t_i | X = x_i), if t*_i < c_i and tau }

The function uses the \code{\link[pec:ipcw]{pec::ipcw()}} \insertCite{pec}{IPCWJK} function to
compute the IPCW weights. They are not normalized (do \strong{not} sum
to one).
}
\examples{
data <- data.frame(
  t = c(5, 8, 12, 15, 20),
  delta = c(1, 0, 1, 0, 1)
)
w <- ipcw_weights(data, tau = 10)
}
\references{
\insertAllCited{}
}
\seealso{
\link{IPCWJK} for more information and \link{ipcwmodel} for implementations
of ready to use models.
}
