% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0IPCWJK-package.R
\docType{package}
\name{IPCWJK-package}
\alias{IPCWJK}
\alias{IPCWJK-package}
\title{IPCWJK: IPCW Jackknife SE Estimation for right-censored survival data
\loadmathjax}
\description{
Provides functions for fitting binary classification models with
inverse probability of censoring weights (IPCW) to estimate survival
probabilities, and implements Jackknife resampling methods for unbiased
prediction error estimation in survival analysis. The package supports model
fitting, IPCW calculation, and Jackknife error estimation.
}
\references{
\insertAllCited{}
}
\author{
\strong{Maintainer}: Lukas Klein \email{lukas.klein@h-da.de}

Authors:
\itemize{
  \item Antje Jahn-Eimermacher
}

}
