% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipcwmodel.R
\name{ipcwmodel}
\alias{ipcwmodel}
\alias{predict.ipcwmodel}
\title{IPCW Model Class}
\usage{
ipcwmodel(
  model_name,
  full_model,
  jackknife_models,
  tau,
  predict,
  train_brier,
  time_var,
  status_var,
  training_vars,
  w,
  additional_information = list()
)

\method{predict}{ipcwmodel}(object, newdata, naive = FALSE, wald = FALSE, z = 1.96, ...)
}
\arguments{
\item{model_name}{Character. The name of the model.}

\item{full_model}{The fitted model object for the full dataset.}

\item{jackknife_models}{List of fitted models, each omitting one observation
(the jackknife resamples).}

\item{tau}{Numeric. The time horizon at which survival is estimated.}

\item{predict}{Function. The prediction function for the fitted model.}

\item{train_brier}{Numeric. The Brier score for the fitted model on the
training data.}

\item{time_var}{Character. The name of the time variable in the data.}

\item{status_var}{Character. The name of the status variable in the data.}

\item{training_vars}{Character vector. The names of the covariates used for
model fitting.}

\item{w}{Numeric vector. The IPCW weights used for model fitting.}

\item{additional_information}{List. Additional information to be stored in
the model object (default is an empty list).}

\item{object}{An object of class \code{ipcwmodel}.}

\item{newdata}{A data frame containing the covariates for which predictions
are to be made.}

\item{naive}{Logical. If \code{TRUE}, use the naive jackknife variance
estimator. If \code{FALSE}, use the IPCW-weighted estimator.}

\item{wald}{Logical. If \code{TRUE}, compute confidence intervals using the
Wald approach (normal approximation). If \code{FALSE}, use weighted
quantiles of the jackknife predictions.}

\item{z}{Numeric. The z-score to use for the confidence interval.
Default is 1.96, corresponding to a 95\% confidence interval. For the
non-Wald approach, this is is converted to quantiles of the
empirical distribution of the jackknife predictions.}

\item{...}{Additional arguments (currently ignored).}
}
\value{
An object of class \code{ipcwmodel} containing the model details.
}
\description{
\loadmathjax
The resulting object contains the fitted model, jackknife resamples,
prediction function, training Brier score, and metadata about the model.
}
\details{
Constructs an object of class \code{ipcwmodel}.

Models can be used by calling the \code{predict(model, newdata)} method.
It returns a data frame with columns for the prediction, lower and upper
confidence intervals, and standard error.

The confidence intervals and standard errors can be computed using either
the naive approach (using the number of jackknife models minus one as the
denominator) or the robust approach (using the IPCW weights). This
is controlled by the \code{naive} argument in the \code{predict} method.

The confidence intervals then can either be computed using the Wald
approach (using a z-score, default is 1.96 for 95\% CI) or using quantiles
of the jackknife predictions (default). This is controlled by the \code{wald}
argument in the \code{predict} method. The quantiles are computed using
\code{Hmisc::wtd.quantile} function, which allows for weighted quantiles.
The weights are either the IPCW weights or the naive weights, depending on
the \code{naive} argument.
}
\section{Methods (by generic)}{
\itemize{
\item \code{predict(ipcwmodel)}: Predict method for \code{ipcwmodel} objects.

}}
\references{
\insertAllCited{}
}
\concept{ipcwbaseclass}
